requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==3.0.0
langchain==0.1.2
flask==3.0.3
pypdf==4.3.1
python-dotenv==1.0.1
-e .
langchain-pinecone==0.0.3
langchain-community==0.0.18
langsmith==0.0.83
langchain-core==0.1.19